修bug

+++ src/main/java/com/ts/cpfr/service/impl/DeviceServiceImpl.java
@@ -161,15 +161,17 @@ public class DeviceServiceImpl implements DeviceService {
     @Override
     public ResultData<ParamData> deleteDevice(ParamData pd) throws Exception {
         String deviceSn = mDeviceDao.selectDeviceSnByDeviceID(pd);
-        pd.put("device_sn", deviceSn);
-        if (mDeviceDao.deleteDeviceByDeviceID(pd)) {
-            TextMessage message = mSocketMessageHandle.obtainMessage(SocketEnum.CODE_1005_DEVICE_DELETE, null);
-            mSocketMessageHandle.sendMessageToDevice(deviceSn, message);
-            //移除app缓存
-            appMemory.removeCache(deviceSn);
-            return new ResultData<>(HandleEnum.SUCCESS);
-        } else
-            return new ResultData<>(HandleEnum.FAIL);
+        if (!StringUtils.isEmpty(deviceSn)) {
+            pd.put("device_sn", deviceSn);
+            if (mDeviceDao.deleteDeviceByDeviceID(pd)) {
+                TextMessage message = mSocketMessageHandle.obtainMessage(SocketEnum.CODE_1005_DEVICE_DELETE, null);
+                mSocketMessageHandle.sendMessageToDevice(deviceSn, message);
+                //移除app缓存
+                appMemory.removeCache(deviceSn);
+                return new ResultData<>(HandleEnum.SUCCESS);
+            }
+        }
+        return new ResultData<>(HandleEnum.FAIL);
     }
 
     @Override
@@ -271,7 +273,7 @@ public class DeviceServiceImpl implements DeviceService {
 
         List deviceIds = CommUtil.getIntListFromObjList(gDeviceList, "device_id");
 
-        if (deviceIds != null){
+        if (deviceIds != null) {
             Iterator<ParamData> it = deviceList.iterator();
             while (it.hasNext()) {
                 ParamData device = it.next();
